pyproject.toml
[tool.poetry]
name = "elevenlabs_tts"
version = "1.0.0"
description = "A Custom Text To Speech integration using ElevenLabs API"
authors = ["@carleeno"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.4"


[tool.poetry.group.dev.dependencies]
homeassistant = ">=2022.11.0"
pytest-homeassistant-custom-component = ">=0.13.1"
bandit = ">=1.7.0"
black = {version = ">=21.12b0", allow-prereleases = true}
mypy = ">=0.812"
pre-commit = ">=2.11.1"
pydocstyle = ">=6.0.0"
prospector = {extras = ["with_all"], version = ">=1.3.1"}
aiohttp_cors = ">=0.7.0"
pytest-asyncio = ">=0.20.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable='pyproject.toml:version,custom_components/elevenlabs_tts/const.py:VERSION'
version_pattern='custom_components/elevenlabs_tts/manifest.json:"version": "{version}"'
upload_to_pypi=false
upload_to_release=false
changelog_sections="feature,fix,breaking,documentation,performance,refactor"
build_command = "echo"
branch = "main"
commit_subject="[skip ci] {version}"

[tool.black]
line-length = 88
target-version = ['py310']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | deps
                     # the root of the project
)
'''

[tool.pytest.ini_options]
minversion = "7.2"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "auto"